third_party/WebKit/LayoutTests/media/video-dom-src.html

<!DOCTYPE html>
<title>Test media "src" attribute set via DOM.</title>
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<video></video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    assert_equals(video.currentSrc, "");
    var mediaFile = "content/test.ogv";
    video.src = mediaFile;
    assert_equals(video.currentSrc, "");

    video.oncanplaythrough = t.step_func_done(function() {
        var currentSrc = video.currentSrc;
        assert_equals(currentSrc.substr(currentSrc.lastIndexOf("/media/")+7), mediaFile);
    });
});
</script>
